Update proc.c -modify proc.c

+++ xv6-public/proc.c
@@ -6,6 +6,7 @@
 #include "x86.h"
 #include "proc.h"
 #include "spinlock.h"
+#include "stddef.h"
 
 struct {
   struct spinlock lock;
@@ -708,20 +709,20 @@ set_cpu_share(int percent){
 	struct  proc *p;
 	int i;
 	int j;
-	if(mlfq_ticket - int(max_tickets*percent/100)<=200)
+	if(mlfq_ticket - (max_tickets*percent/100) <= 200)
 		return -1;
 	// take tickets
 	else{
 		//stide queue에 프로세스를 할당
 		p=myproc();
-		p->tickets=int(max_tickets*percent/100);
+		p->tickets=(max_tickets*percent/100);
 		p->stide=max_tickets/p->tickets;
 		p->pass=min_pass;
 		stride_count++;
 		stride_queue[stride_count]=p;
 		// mlfq의 티켓 변동사항 반영
-		mlfq_ticket=mlfq_ticket - int(max_tickets*percent/100);
-		mlfq_stride=int(max_tickets/mlfq_ticket);
+		mlfq_ticket=mlfq_ticket - (max_tickets*percent/100);
+		mlfq_stride=(max_tickets/mlfq_ticket);
 		// mlfq에서 프로세스 제거
 		for(i=0;i<=high_count;i++)
 		{
